\name{debias-package}
\alias{debias}
\docType{package}
\title{
 Bias Reduction of Small Sample MLE and Related Functions
}
\description{
 Functions implementing MLE,its bias reduction for small samples and related issues as documented
 in The New Weibull Handbook, by Dr. Robert B. Abernethy
}

\section{Introduction}{
	\pkg{debias}  This package is intended to expose the specific implementation of 
	Dr. Abernethy's methods through open source code.  The primary purpose is to encourage 
	further work in the feild of reliability analysis initiated by Waloddi Weibull, and 
	continued by Dr. Robert B. Abernethy, Paul Barringer, Wes Fulton and many others.
	Ultimately, this package may progress to permit development of a complete and bonifide 
	user application.
}

\author{
Jacob T. Ormerod

Maintainer: David Silkworth <davidsilkworth@openreliability.me>
}                               
\references{
  Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"

}

\keyword{ package }



